backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.37
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.58.0
apscheduler==3.10.4
sqlalchemy==2.0.47
python-slugify==8.0.4
python-dateutil==2.9.0
python-multipart==0.0.20
